Chore: 댓글, 대댓글 생성 시 LazyInitializationException에러 해결

+++ beer_community/src/main/java/team_project/beer_community/dto/WriteCommentDto.java
@@ -11,12 +11,12 @@ public class WriteCommentDto {
     @NotNull
     private String content;
     @NotNull
-    private double point;
+    private int point;
 
     @NotNull
     private Long beerId;
 
-    public WriteCommentDto(String content, double point, Long beerId) {
+    public WriteCommentDto(String content, int point, Long beerId) {
         this.content = content;
         this.point = point;
         this.beerId = beerId;
